Set larger threshold for day/night cycle trigger
This fixes a bug wherein adjusting planet parameters in the GUI during
nighttime triggers daytime because the activation distance for the
day/night cycles is 200 pixels from the left and right screen edges. By
setting the threshold to 10 pixels we avoid this problem. This now
requires the cursor to be closer to the left/right screen edges to
trigger a day/night change.